Product Switcher (#239)
Adds a product switcher menu to the sidebar. 

I did it using `headless-ui`, since we don't have a component library
here and I didn't want to bring in one. We already use headlessui in
www.

The customization is done using the `swizzling` feature from Docusaurus,
which lets us use custom components for the documentation layout. See
https://docusaurus.io/docs/swizzling
Note that swizzling requires basically copy-pasting components from the
base Docusaurus theme and changing them. The stuff under `DocsSidebar`
is copied from the base theme, only `ProductSwitcher` was written by me.

Maybe we can add a bit more to it and make it more flashy, but this is a
good start :)
  


https://github.com/user-attachments/assets/a3a63bfb-409b-4b4c-acf8-0149417b4cbb


Closes https://github.com/svix/monorepo-private/issues/11538