Add button to copy docs for llm (#244)
Adds 'Copy docs for LLM' button and add it to a whole bunch of pages.
The button just takes the raw markdown and puts it in the clipboard.
Also has segment tracking to measure how many people use it.

I couldn't add it to the docusaurus theme/layout because docusaurus
doesn't have a good way to edit the header, so it has to be added page
by page. Maybe that's a good thing anyway, because it doesn't make sense
to have it in every single page.

Tom wants this in most pages, so I added it where I think it (even
barely) made sense: docs with code snippets or explaining Svix concepts
that might be useful for an LLM context.


https://github.com/user-attachments/assets/b0da3a87-719a-428e-ae59-4fded65c35aa

Closes https://github.com/svix/monorepo-private/issues/11722